# docs/swagger.yaml
basePath: /
definitions:
  system.Button:
    properties:
      code:
        type: string
      desc:
        type: string
      id:
        type: integer
    type: object
  system.Menu:
    properties:
      activeMenu:
        description: 激活的菜单名
        type: string
      buttons:
        items:
          $ref: '#/definitions/system.Button'
        type: array
      children:
        items:
          $ref: '#/definitions/system.Menu'
        type: array
      component:
        type: string
      constant:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        type: string
      fixedIndexInTab:
        type: integer
      hideInMenu:
        description: 是否隐藏菜单
        type: boolean
      href:
        type: string
      i18nKey:
        type: string
      icon:
        type: string
      iconType:
        type: string
      id:
        type: integer
      keepAlive:
        type: boolean
      localIcon:
        type: string
      menuName:
        type: string
      menuType:
        type: string
      multiTab:
        description: 是否支持多标签
        type: boolean
      order:
        type: integer
      parentId:
        description: 父菜单ID
        type: integer
      query:
        items:
          $ref: '#/definitions/system.MenuQuery'
        type: array
      routeName:
        type: string
      routePath:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  system.MenuQuery:
    properties:
      id:
        type: integer
      key:
        type: string
      menuId:
        type: integer
      value:
        type: string
    type: object
  system.Role:
    properties:
      buttons:
        items:
          $ref: '#/definitions/system.Button'
        type: array
      createTime:
        type: string
      deleteTime:
        type: string
      id:
        type: integer
      menu:
        items:
          $ref: '#/definitions/system.Menu'
        type: array
      roleDesc:
        type: string
      roleName:
        type: string
      updateTime:
        type: string
    required:
    - roleName
    type: object
...
  utils.Pagination-array_system_Role:
    properties:
      current:
        type: integer
      records:
        items:
          items:
            $ref: '#/definitions/system.Role'
          type: array
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
...
  utils.Response-array_system_Role:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/system.Role'
        type: array
      msg:
        type: string
    type: object
  utils.Response-string:
    properties:
      code:
        type: string
      data:
        type: string
      msg:
        type: string
    type: object
  utils.Response-system_Role:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/system.Role'
      msg:
        type: string
    type: object
...
  utils.Response-utils_Pagination-array_system_Role:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/utils.Pagination-array_system_Role'
      msg:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Soybean Admin Go 后台管理系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Soybean Admin Go API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: 获取分页的角色列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-utils_Pagination-array_system_Role'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 获取分页角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建新角色
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/system.Role'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-system_Role'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 创建角色
      tags:
      - 角色管理
  /roles/all:
    get:
      consumes:
      - application/json
      description: 获取所有角色列表（不分页）
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-array_system_Role'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 获取所有角色
      tags:
      - 角色管理
  /roles/batchDelete:
    post:
      consumes:
      - application/json
      description: 批量删除多个角色
      parameters:
      - description: 角色ID列表
        in: body
        name: ids
        required: true
        schema:
          properties:
            ids:
              items:
                type: integer
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-string'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 批量删除角色
      tags:
      - 角色管理
  /roles/buttons:
    post:
      consumes:
      - application/json
      description: 更新角色的按钮权限
      parameters:
      - description: 角色按钮信息
        in: body
        name: buttonInfo
        required: true
        schema:
          properties:
            buttonIds:
              items:
                type: integer
              type: array
            roleId:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-string'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 更新角色按钮权限
      tags:
      - 角色管理
  /roles/delete:
    post:
      consumes:
      - application/json
      description: 删除指定角色
      parameters:
      - description: 角色ID
        in: body
        name: id
        required: true
        schema:
          properties:
            id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-string'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 删除角色
      tags:
      - 角色管理
  /roles/menus:
    post:
      consumes:
      - application/json
      description: 更新角色的菜单权限
      parameters:
      - description: 角色菜单信息
        in: body
        name: menuInfo
        required: true
        schema:
          properties:
            menuIds:
              items:
                type: integer
              type: array
            roleId:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-string'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 更新角色菜单
      tags:
      - 角色管理
  /roles/update:
    post:
      consumes:
      - application/json
      description: 更新角色信息
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/system.Role'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response-string'
        "400":
          description: 错误
          schema:
            $ref: '#/definitions/utils.Response-string'
      security:
      - ApiKeyAuth: []
      summary: 更新角色
      tags:
      - 角色管理
...
securityDefinitions:
  ApiKeyAuth:
    description: '请在此处输入Bearer Token，格式为: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"